requirements.txt
kafka-python==1.2.5
six==1.10.0
